Final refactor of articles lib - resulting in only article-related functions in the lib, yay
