+++ lib/articles.js
@@ -856,316 +856,6 @@ export async function hasuraInsertArticleSlugVersions(params) {
   });
 }
 
-const HASURA_LOOKUP_GOOGLE_DOC = `query ApiLookupGoogleDoc($document_id: String!) {
-  google_documents(where: {document_id: {_eq: $document_id}}) {
-    article_google_documents {
-      article {
-        id
-        slug
-        canonical_url
-        category {
-          id
-          slug
-          title
-        }
-        author_articles {
-          author {
-            id
-            first_names
-            last_name
-            slug
-          }
-        }
-        article_sources {
-          source {
-            affiliation
-            age
-            created_at
-            email
-            ethnicity
-            gender
-            id
-            name
-            phone
-            race
-            role
-            sexual_orientation
-            updated_at
-            zip
-          }
-        }
-        tag_articles {
-          tag_id
-        }
-        article_translations(order_by: {id: desc}, limit: 1) {
-          custom_byline
-          facebook_description
-          facebook_title
-          first_published_at
-          headline
-          id
-          last_published_at
-          locale_code
-          published
-          search_description
-          search_title
-          twitter_description
-          twitter_title
-          published_article_translations {
-            article_translation {
-              first_published_at
-              id
-              locale_code
-              last_published_at
-            }
-          }
-        }
-      }
-      google_document {
-        locale_code
-      }
-    }
-    page_google_documents {
-      google_document {
-        locale_code
-      }
-      page {
-        id
-        slug
-        author_pages {
-          author {
-            first_names
-            last_name
-            id
-            slug
-          }
-        }
-        page_translations(limit: 1, order_by: {id: desc}) {
-          facebook_description
-          facebook_title
-          first_published_at
-          headline
-          id
-          last_published_at
-          locale_code
-          published
-          search_description
-          search_title
-          twitter_title
-          twitter_description
-        }
-      }
-    }
-  }
-  authors {
-    id
-    slug
-    first_names
-    last_name
-  }
-  categories {
-    id
-    published
-    slug
-    category_translations(where: {locale_code: {_eq: "en-US"}}) {
-      locale_code
-      title
-    }
-  }
-  homepage_layout_datas {
-    article_priority_1
-    article_priority_2
-    article_priority_3
-  }
-  organization_locales {
-    locale {
-      code
-      name
-    }
-  }
-  tags {
-    id
-    slug
-    tag_translations(where: {locale_code: {_eq: "en-US"}}) {
-      locale_code
-      title
-    }
-  }
-}`;
-
-export async function hasuraLookupGoogleDoc(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: HASURA_LOOKUP_GOOGLE_DOC,
-    name: 'ApiLookupGoogleDoc',
-    variables: {
-      document_id: params['documentId'],
-    },
-  });
-}
-
-const HASURA_GET_GOOGLE_DOCS_FOR_ARTICLE = `query ApiLookupGoogleDocsForArticle($article_id: Int!, $locale_code: String!) {
-  article_google_documents(where: {article_id: {_eq: $article_id}}) {
-    google_document {
-      document_id
-      locale_code
-      locale {
-        code
-        name
-      }
-      url
-    }
-    article_id
-  }
-  published_article_translations(where: {locale_code: {_eq: $locale_code}, article_id: {_eq: $article_id}}) {
-    article_translation {
-      id
-      first_published_at
-      last_published_at
-      locale_code
-    }
-  }
-}`;
-
-export async function hasuraGetGoogleDocsForArticle(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: HASURA_GET_GOOGLE_DOCS_FOR_ARTICLE,
-    name: 'ApiLookupGoogleDocsForArticle',
-    variables: {
-      article_id: params['articleId'],
-      locale_code: params['localeCode'],
-    },
-  });
-}
-
-const HASURA_GET_GOOGLE_DOCS_FOR_PAGE = `query ApiLookupGoogleDocsForPage($page_id: Int!) {
-  page_google_documents(where: {page_id: {_eq: $page_id}}) {
-    google_document {
-      document_id
-      locale_code
-      locale {
-        code
-        name
-      }
-      url
-    }
-    page_id
-  }
-}`;
-
-export async function hasuraGetGoogleDocsForPage(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    site: params['site'],
-    query: HASURA_GET_GOOGLE_DOCS_FOR_PAGE,
-    name: 'ApiLookupGoogleDocsForPage',
-    variables: {
-      page_id: params['pageId'],
-    },
-  });
-}
-
-const HASURA_LIST_STATIC_PREVIEW_PAGE_PATHS = `query FrontendListStaticPreviewPagePaths {
-  organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
-    customDomain
-    subdomain
-    organization_locales {
-      locale {
-        code
-      }
-    }
-    pages {
-      slug
-      page_translations(distinct_on: locale_code) {
-        locale_code
-        published
-        headline
-      }
-    }
-  }
-}`;
-
-export async function listStaticPreviewPagePaths(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    adminSecret: params['adminSecret'],
-    query: HASURA_LIST_STATIC_PREVIEW_PAGE_PATHS,
-    name: 'FrontendListStaticPreviewPagePaths',
-  });
-}
-
-const HASURA_LIST_STATIC_PAGE_PATHS = `query FrontendListStaticPagePaths {
-  organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
-    customDomain
-    subdomain
-    organization_locales {
-      locale {
-        code
-      }
-    }
-    pages(where: {page_translations: {published: {_eq: true}}}) {
-      slug
-      page_translations(where: {published: {_eq: true}}, distinct_on: locale_code) {
-        locale_code
-        published
-        headline
-      }
-    }
-  }
-}`;
-
-export async function listStaticPagePaths(params) {
-  return fetchGraphQL({
-    url: params['url'],
-    adminSecret: params['adminSecret'],
-    query: HASURA_LIST_STATIC_PAGE_PATHS,
-    name: 'FrontendListStaticPagePaths',
-  });
-}
-
-export async function generateAllStaticPagePaths(params = {}) {
-  let result;
-  if (params['preview']) {
-    result = await listStaticPreviewPagePaths(params);
-  } else {
-    result = await listStaticPagePaths(params);
-  }
-
-  if (result.errors) {
-    throw result.errors;
-  }
-
-  let allPaths = [];
-  for (let i = 0; i < result.data.organizations.length; i++) {
-    let org = result.data.organizations[i];
-
-    for (let x = 0; x < org.pages.length; x++) {
-      let page = org.pages[x];
-      let siteParams = Object.assign({}, params['urlParams']);
-      if (org.subdomain) {
-        siteParams['site'] = org.subdomain;
-      } else if (org.customDomain) {
-        siteParams['site'] = org.customDomain;
-      }
-      siteParams['slug'] = page.slug;
-      allPaths.push({
-        params: siteParams,
-      });
-      // for (let y = 0; y < page.page_translations.length; y++) {
-      //   allPaths.push({
-      //     params: siteParams,
-      //     locale: page.page_translations[y].locale_code,
-      //   });
-      // }
-    }
-  }
-
-  return allPaths;
-}
-
 const HASURA_LIST_ALL_ARTICLE_ARCHIVE_PAGES = `query FrontendPaginatedListArticlesArchivePages {
   organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
     customDomain
@@ -1195,6 +885,7 @@ export async function listArticleArchivePages(params) {
 function range(size, startAt = 0) {
   return [...Array(size).keys()].map((i) => i + startAt);
 }
+
 export async function generateAllArticlePagePaths(params = {}) {
   const { errors, data } = await hasuraListAllArticleSlugs();
   if (errors) {
