+++ lib/document.js
@@ -3,6 +3,218 @@ import { fetchGraphQL, slugify } from './graphql';
 import { getFacebookMarkup } from './embeds';
 import TinyS3 from './tiny_s3';
 
+const HASURA_LOOKUP_GOOGLE_DOC = `query ApiLookupGoogleDoc($document_id: String!) {
+  google_documents(where: {document_id: {_eq: $document_id}}) {
+    article_google_documents {
+      article {
+        id
+        slug
+        canonical_url
+        category {
+          id
+          slug
+          title
+        }
+        author_articles {
+          author {
+            id
+            first_names
+            last_name
+            slug
+          }
+        }
+        article_sources {
+          source {
+            affiliation
+            age
+            created_at
+            email
+            ethnicity
+            gender
+            id
+            name
+            phone
+            race
+            role
+            sexual_orientation
+            updated_at
+            zip
+          }
+        }
+        tag_articles {
+          tag_id
+        }
+        article_translations(order_by: {id: desc}, limit: 1) {
+          custom_byline
+          facebook_description
+          facebook_title
+          first_published_at
+          headline
+          id
+          last_published_at
+          locale_code
+          published
+          search_description
+          search_title
+          twitter_description
+          twitter_title
+          published_article_translations {
+            article_translation {
+              first_published_at
+              id
+              locale_code
+              last_published_at
+            }
+          }
+        }
+      }
+      google_document {
+        locale_code
+      }
+    }
+    page_google_documents {
+      google_document {
+        locale_code
+      }
+      page {
+        id
+        slug
+        author_pages {
+          author {
+            first_names
+            last_name
+            id
+            slug
+          }
+        }
+        page_translations(limit: 1, order_by: {id: desc}) {
+          facebook_description
+          facebook_title
+          first_published_at
+          headline
+          id
+          last_published_at
+          locale_code
+          published
+          search_description
+          search_title
+          twitter_title
+          twitter_description
+        }
+      }
+    }
+  }
+  authors {
+    id
+    slug
+    first_names
+    last_name
+  }
+  categories {
+    id
+    published
+    slug
+    category_translations(where: {locale_code: {_eq: "en-US"}}) {
+      locale_code
+      title
+    }
+  }
+  homepage_layout_datas {
+    article_priority_1
+    article_priority_2
+    article_priority_3
+  }
+  organization_locales {
+    locale {
+      code
+      name
+    }
+  }
+  tags {
+    id
+    slug
+    tag_translations(where: {locale_code: {_eq: "en-US"}}) {
+      locale_code
+      title
+    }
+  }
+}`;
+
+export async function hasuraLookupGoogleDoc(params) {
+  return fetchGraphQL({
+    url: params['url'],
+    site: params['site'],
+    query: HASURA_LOOKUP_GOOGLE_DOC,
+    name: 'ApiLookupGoogleDoc',
+    variables: {
+      document_id: params['documentId'],
+    },
+  });
+}
+
+const HASURA_GET_GOOGLE_DOCS_FOR_ARTICLE = `query ApiLookupGoogleDocsForArticle($article_id: Int!, $locale_code: String!) {
+  article_google_documents(where: {article_id: {_eq: $article_id}}) {
+    google_document {
+      document_id
+      locale_code
+      locale {
+        code
+        name
+      }
+      url
+    }
+    article_id
+  }
+  published_article_translations(where: {locale_code: {_eq: $locale_code}, article_id: {_eq: $article_id}}) {
+    article_translation {
+      id
+      first_published_at
+      last_published_at
+      locale_code
+    }
+  }
+}`;
+
+export async function hasuraGetGoogleDocsForArticle(params) {
+  return fetchGraphQL({
+    url: params['url'],
+    site: params['site'],
+    query: HASURA_GET_GOOGLE_DOCS_FOR_ARTICLE,
+    name: 'ApiLookupGoogleDocsForArticle',
+    variables: {
+      article_id: params['articleId'],
+      locale_code: params['localeCode'],
+    },
+  });
+}
+
+const HASURA_GET_GOOGLE_DOCS_FOR_PAGE = `query ApiLookupGoogleDocsForPage($page_id: Int!) {
+  page_google_documents(where: {page_id: {_eq: $page_id}}) {
+    google_document {
+      document_id
+      locale_code
+      locale {
+        code
+        name
+      }
+      url
+    }
+    page_id
+  }
+}`;
+
+export async function hasuraGetGoogleDocsForPage(params) {
+  return fetchGraphQL({
+    url: params['url'],
+    site: params['site'],
+    query: HASURA_GET_GOOGLE_DOCS_FOR_PAGE,
+    name: 'ApiLookupGoogleDocsForPage',
+    variables: {
+      page_id: params['pageId'],
+    },
+  });
+}
+
 export async function processDocumentContents(
   elements,
   listInfo,
