+++ lib/pages.js
@@ -286,6 +286,26 @@ const HASURA_CREATE_PAGE = `mutation FrontendTestingCreatePage($slug: String!, $
   }
 }`;
 
+const HASURA_LIST_STATIC_PREVIEW_PAGE_PATHS = `query FrontendListStaticPreviewPagePaths {
+  organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
+    customDomain
+    subdomain
+    organization_locales {
+      locale {
+        code
+      }
+    }
+    pages {
+      slug
+      page_translations(distinct_on: locale_code) {
+        locale_code
+        published
+        headline
+      }
+    }
+  }
+}`;
+
 export async function hasuraCreatePage(params) {
   let url = params['url'] || process.env.HASURA_API_URL;
   let orgSlug = params['orgSlug'] || process.env.ORG_SLUG;
@@ -306,4 +326,82 @@ export async function hasuraCreatePage(params) {
       created_by_email: params['created_by_email'],
     },
   });
+}
+
+export async function listStaticPreviewPagePaths(params) {
+  return fetchGraphQL({
+    url: params['url'],
+    adminSecret: params['adminSecret'],
+    query: HASURA_LIST_STATIC_PREVIEW_PAGE_PATHS,
+    name: 'FrontendListStaticPreviewPagePaths',
+  });
+}
+
+const HASURA_LIST_STATIC_PAGE_PATHS = `query FrontendListStaticPagePaths {
+  organizations(where: {_or: [{subdomain: {_is_null: false}}, {customDomain: {_is_null: false}}]}) {
+    customDomain
+    subdomain
+    organization_locales {
+      locale {
+        code
+      }
+    }
+    pages(where: {page_translations: {published: {_eq: true}}}) {
+      slug
+      page_translations(where: {published: {_eq: true}}, distinct_on: locale_code) {
+        locale_code
+        published
+        headline
+      }
+    }
+  }
+}`;
+
+export async function listStaticPagePaths(params) {
+  return fetchGraphQL({
+    url: params['url'],
+    adminSecret: params['adminSecret'],
+    query: HASURA_LIST_STATIC_PAGE_PATHS,
+    name: 'FrontendListStaticPagePaths',
+  });
+}
+
+export async function generateAllStaticPagePaths(params = {}) {
+  let result;
+  if (params['preview']) {
+    result = await listStaticPreviewPagePaths(params);
+  } else {
+    result = await listStaticPagePaths(params);
+  }
+
+  if (result.errors) {
+    throw result.errors;
+  }
+
+  let allPaths = [];
+  for (let i = 0; i < result.data.organizations.length; i++) {
+    let org = result.data.organizations[i];
+
+    for (let x = 0; x < org.pages.length; x++) {
+      let page = org.pages[x];
+      let siteParams = Object.assign({}, params['urlParams']);
+      if (org.subdomain) {
+        siteParams['site'] = org.subdomain;
+      } else if (org.customDomain) {
+        siteParams['site'] = org.customDomain;
+      }
+      siteParams['slug'] = page.slug;
+      allPaths.push({
+        params: siteParams,
+      });
+      // for (let y = 0; y < page.page_translations.length; y++) {
+      //   allPaths.push({
+      //     params: siteParams,
+      //     locale: page.page_translations[y].locale_code,
+      //   });
+      // }
+    }
+  }
+
+  return allPaths;
 }
